requirements.txt
# Dependencies for Laravel Developer Agent
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-anthropic>=0.1.0
langgraph>=0.0.20
anthropic>=0.8.0
python-dotenv>=1.0.0
typer[all]>=0.9.0
rich>=12.0.0
colorama>=0.4.6
pydantic>=2.4.0
jinja2>=3.1.2
requests>=2.28.0 